Adds apps.gallery.utils with create_test_data.py function.
